Have floats near zero converge to 0.000001
If you have a fuzz test that fails for all positive or
all negative floats, shrinking will produce one answer instead
of many small floats.

This value was chosen because it is the largest positive float
not to be toString-ed into scientific notation.

My concern is that this feels like special casing, because if
your test fails with values greater than 1, it will not converge
onto 1.000001.